refactor(app): extract the query runners
The buffered runner and the streaming "No limit" runner move to
`wire::runner`. It exposes `build` rather than `wire`: it installs no
callbacks, it constructs the two closures every query-triggering handler
calls, and `main` hands them on through `AppFns`.

With this `fn main` is down to roughly 740 lines from 8,300: build the
state, build the shared closures, call the wiring modules, run the event
loop.
