+++ app/src/wire/mod.rs
@@ -13,6 +13,7 @@ pub(crate) mod find;
 pub(crate) mod grid;
 pub(crate) mod picker;
 pub(crate) mod query;
+pub(crate) mod runner;
 pub(crate) mod schema;
 pub(crate) mod settings;
 pub(crate) mod split_pane;
